refactor: load tests from src by giving src real .ts import extensions
Following PR review: the worker/fixture/stress tests imported the built dist,
which required a bundle build before running them. Convert src's relative
imports to real .ts extensions (the form our TypeStrip convention prescribes)
so Node's native type stripping can load src directly, then point the
native-loaded test files back at ../../src/index.ts.

- src/*.ts: 56 relative imports .js -> .ts (incl. one side-effect import)
- test/workers, test/fixtures, stress-test/workers: import src/index.ts, not dist
- vitest{,-stress}.config.ts: version banner imports src via a computed
  specifier so the config bundler no longer hard-requires dist to exist; it
  degrades to '?' when the module can't load
- benchmarks keep importing dist on purpose (they measure the shipped artifact;
  bench workflow builds first)
- AGENTS.md: document the native-strip rules (real extensions, verbatim type
  imports, src-loaded tests need no bundle build)

Worker/fixture/stress tests now run with no dist present; verified the full
suite (756 passed), benchmarks, stress, type-check, lint, and format.
