+++ src/index.ts
@@ -1,4 +1,4 @@
-import { version } from './load-binding.js';
+import { version } from './load-binding.ts';
 
 export {
 	backups,
@@ -6,18 +6,18 @@ export {
 	type BackupOptions,
 	type RestoreMode,
 	type RestoreOptions,
-} from './backup.js';
-export type { BackupStreamOptions } from './backup-stream.js';
+} from './backup.ts';
+export type { BackupStreamOptions } from './backup-stream.ts';
 export {
 	RocksDatabase,
 	type RocksDatabaseOptions,
 	type RocksDBStat,
 	type RocksDBStats,
-} from './database.js';
-export { DBIterator } from './dbi-iterator.js';
-export { DBI, type IteratorOptions } from './dbi.js';
-export type { Key } from './encoding.js';
-export type * from './stats.js';
+} from './database.ts';
+export { DBIterator } from './dbi-iterator.ts';
+export { DBI, type IteratorOptions } from './dbi.ts';
+export type { Key } from './encoding.ts';
+export type * from './stats.ts';
 export {
 	constants,
 	coolTransactionLogs,
@@ -32,14 +32,14 @@ export {
 	type TransactionEntry,
 	type TransactionLogPosition,
 	type TransactionLogStats,
-} from './load-binding.js';
-export * from './parse-transaction-log.js';
+} from './load-binding.ts';
+export * from './parse-transaction-log.ts';
 export {
 	validateTransactionLogStore,
 	type TransactionLogFileValidation,
 	type TransactionLogStoreValidation,
 	type ValidateTransactionLogStoreOptions,
-} from './validate-transaction-log.js';
+} from './validate-transaction-log.ts';
 export {
 	type CompressionAlgorithm,
 	type CompressionInfo,
@@ -52,11 +52,11 @@ export {
 	type StorePutOptions,
 	type StoreRangeOptions,
 	type StoreRemoveOptions,
-} from './store.js';
-export { Transaction } from './transaction.js';
+} from './store.ts';
+export { Transaction } from './transaction.ts';
 
-import './transaction-log-reader.js'; // installs TransactionLog.prototype.query
-export { CorruptFrameError } from './transaction-log-reader.js';
+import './transaction-log-reader.ts'; // installs TransactionLog.prototype.query
+export { CorruptFrameError } from './transaction-log-reader.ts';
 
 export const versions: { rocksdb: string; 'rocksdb-js': string } = {
 	rocksdb: version,
