+++ stress-test/workers/stress-transaction-log-mmap-worker.mts
@@ -1,4 +1,4 @@
-import { RocksDatabase } from '../../dist/index.mjs';
+import { RocksDatabase } from '../../src/index.ts';
 import { setImmediate as yieldTick } from 'node:timers/promises';
 import { parentPort, workerData } from 'node:worker_threads';
