+++ test/fixtures/fork-close-commit-uaf.mts
@@ -29,7 +29,7 @@
  *
  * Exit 0 = survived; a crash exits via signal / non-zero.
  */
-import { RocksDatabase } from '../../dist/index.mjs';
+import { RocksDatabase } from '../../src/index.ts';
 import { createWorkerBootstrapScript } from '../lib/worker-bootstrap.ts';
 import { mkdirSync } from 'node:fs';
 import { Worker } from 'node:worker_threads';
