+++ test/fixtures/fork-commit-teardown.mts
@@ -28,7 +28,7 @@
  * Set ROCKSDB_JS_COMMIT_DELAY_MS (see the test) to widen the window so the
  * race reproduces deterministically.
  */
-import { RocksDatabase } from '../../dist/index.mjs';
+import { RocksDatabase } from '../../src/index.ts';
 import { createWorkerBootstrapScript } from '../lib/worker-bootstrap.ts';
 import { mkdirSync } from 'node:fs';
 import { setTimeout as delay } from 'node:timers/promises';
