+++ test/fixtures/fork-notify-teardown-uaf.mts
@@ -23,7 +23,7 @@
  * this isolates the notify path from the unrelated transaction-close double
  * free. Exit 0 = survived; a crash exits via signal / non-zero.
  */
-import { RocksDatabase } from '../../dist/index.mjs';
+import { RocksDatabase } from '../../src/index.ts';
 import { createWorkerBootstrapScript } from '../lib/worker-bootstrap.ts';
 import { mkdirSync } from 'node:fs';
 import { setTimeout as delay } from 'node:timers/promises';
