+++ test/workers/commit-teardown-worker.mts
@@ -1,4 +1,4 @@
-import { RocksDatabase } from '../../dist/index.mjs';
+import { RocksDatabase } from '../../src/index.ts';
 import { parentPort, workerData } from 'node:worker_threads';
 
 // Open the SAME path as the parent so the parent's handle pins the shared
