+++ test/workers/global-events-worker.mts
@@ -1,4 +1,4 @@
-import { RocksDatabase } from '../../dist/index.mjs';
+import { RocksDatabase } from '../../src/index.ts';
 import { parentPort } from 'node:worker_threads';
 
 RocksDatabase.on('global-events-test:from-main', (value) => {
