+++ test/workers/thread-id-worker.mts
@@ -1,4 +1,4 @@
-import { currentThreadId } from '../../dist/index.mjs';
+import { currentThreadId } from '../../src/index.ts';
 import { parentPort } from 'node:worker_threads';
 
 parentPort?.postMessage({ threadId: currentThreadId() });
