+++ test/workers/txn-close-commit-worker.mts
@@ -1,4 +1,4 @@
-import { RocksDatabase } from '../../dist/index.mjs';
+import { RocksDatabase } from '../../src/index.ts';
 import { parentPort, workerData } from 'node:worker_threads';
 
 // W opens a FRESH handle each round (signaled by M via "nextRound") so the
