+++ test/workers/user-shared-buffer-uaf-close-worker.mts
@@ -1,4 +1,4 @@
-import { RocksDatabase } from '../../dist/index.mjs';
+import { RocksDatabase } from '../../src/index.ts';
 import { parentPort, workerData } from 'node:worker_threads';
 
 const db = RocksDatabase.open(workerData.path, { name: workerData.name });
